[강운덕] [PINPOINT-267] histogram의 timeslot 값이 int와 short이 혼용되는 부분이 있어 int로 수정함.  - short일때 리턴값 수정.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-web/trunk@3237 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/web/vo/LinkStatistics.java
@@ -161,11 +161,11 @@ public class LinkStatistics {
 		return timeseriesValueList;
 	}
 
-	public int getSlow() {
+	public short getSlow() {
 		return SLOT_SLOW;
 	}
 
-	public int getError() {
+	public short getError() {
 		return SLOT_ERROR;
 	}
